Extract api error helpers
